feat(ui): improve vesting admin alignment and clarity
- Align "Vesting Candidates" and "Active Vesting Schedules" margins
- Rename "Pending Mock Vesting Transactions" → "Vesting Candidates"
- Remove "tokens" suffix from AMOUNT column header
- Sort Active Schedules with "Ready to Release" first
- Sidebar: "Schedule Vesting" → "New Vesting"
- Normalize section headers and alignment
